% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_word.R
\name{create_word}
\alias{create_word}
\title{create Rmd file to generate Word document.
create yaml header for rmd to output word document}
\usage{
create_word()
}
\value{
rmd document
}
\description{
create Rmd file to generate Word document.
create yaml header for rmd to output word document
}
