% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_section.R
\name{insert_section}
\alias{insert_section}
\title{Insert the paragraph of SAP topic. One topic at a time.
Rstudio Addin}
\usage{
insert_section()
}
\value{
Paragraph of the SAP topic
}
\description{
Insert the paragraph of SAP topic. One topic at a time.
Rstudio Addin
}
